Replace console.log with debug
To make it possible to filter out the logs that you’re interested in.
I separated the logs in 2 different namespaces error and log, so that
you can e.g. filter out logs when you are trying to fix an error.

To only show errors you set the DEBUG environment variable to `error`