Added preflight options request endpoint
This was needed for testing purposes since my http request library
always sends a preflight request first.

I’m also adding it to the production environment since it might be
useful to create requests to the server from e.g. the browser for
debugging purposes and then we’ll make preflight requests.